Move files around and get register working + session refresh working
