+++ frontend/package.json
@@ -10,7 +10,7 @@
   },
   "dependencies": {
     "@heroicons/react": "^2.1.4",
-    "bcrypt": "^5.1.1",
+    "bcryptjs": "^2.4.3",
     "clsx": "^2.1.1",
     "jose": "^5.6.3",
     "next": "14.2.3",
@@ -19,8 +19,8 @@
   },
   "devDependencies": {
     "@prisma/client": "^5.15.0",
-    "@types/bcrypt": "^5.0.2",
     "@types/node": "^20.14.2",
+    "@types/bcryptjs": "^2.4.6",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "eslint": "^8",
